feat(semver): Support semver based versions
This will allow semvers to be passed in with the request, e.g.
`/^v1.0.0/blogs`. As a result, consumers can safely hardcode a
sane API version in their codebase but won't miss out on bug fixes
or new features from the API.

The original methods `setByPath` and `setByAccept` still behave the
same but their signatures have changed.

BREAKING CHANGE: This not only uses ES6 features, but also changes
the funciton signature for the exported methods.