feat(semver): Prevent locking to prerelease
If prereleases are exposed to consumers you may wish to prevent them
from being locked in requests. A version range would be safer since
as soon as you release a stable version you consumers would receive
that instead.

This change will allow you to prevent requests for locked prereleases
and instead return a 400 in those cases.